Migration to ASP.NET 5 Of ProvisioningSystem
