--- AzureProvisioningJob/packages.config
@@ -1,29 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Azure.WebJobs" version="1.0.0" targetFramework="net45" />
-  <package id="Microsoft.Azure.WebJobs.Core" version="1.0.0" targetFramework="net45" />
-  <package id="Microsoft.Bcl" version="1.1.9" targetFramework="net45" />
-  <package id="Microsoft.Bcl.Async" version="1.0.168" targetFramework="net45" />
-  <package id="Microsoft.Bcl.Build" version="1.0.14" targetFramework="net45" />
-  <package id="Microsoft.Data.Edm" version="5.6.0" targetFramework="net45" />
-  <package id="Microsoft.Data.OData" version="5.6.0" targetFramework="net45" />
-  <package id="Microsoft.Data.Services.Client" version="5.6.0" targetFramework="net45" />
-  <package id="Microsoft.IdentityModel.Clients.ActiveDirectory" version="2.14.201151115" targetFramework="net45" />
-  <package id="Microsoft.Net.Http" version="2.2.22" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.Common" version="1.3.0" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.Common.Dependencies" version="1.1.0" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.ConfigurationManager" version="2.0.3" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.Management" version="2.0.0" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.Management.Compute" version="5.0.0" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.Management.Libraries" version="2.0.0" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.Management.MediaServices" version="2.0.0" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.Management.Monitoring" version="1.0.0" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.Management.Network" version="3.0.0" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.Management.Scheduler" version="3.0.0" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.Management.Sql" version="3.0.0" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.Management.Storage" version="3.0.0" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.Management.WebSites" version="3.0.0" targetFramework="net45" />
-  <package id="Newtonsoft.Json" version="6.0.4" targetFramework="net45" />
-  <package id="System.Spatial" version="5.6.0" targetFramework="net45" />
-  <package id="WindowsAzure.Storage" version="4.2.1" targetFramework="net45" />
-</packages>
