--- AzureProvisioningLibrary/Configuration.cs
@@ -1,25 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace AzureProvisioningLibrary
-{
-    public static class ProvisioningLibraryConfiguration
-    {
-      
-        public static string GetStorageSubscriptionCertificate()
-        {
-            //This is temporary unitll we get from DB
-            throw new NotImplementedException();
-            return "{add certificate}";
-        }
-        public static string GetStorageSubscriptionId()
-        {
-            //This is temporary unitll we get from DB
-            throw new NotImplementedException();
-            return "{add Subscription Id}";
-        }
-    }
-}
