--- AzureProvisioningLibrary/app.config
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-	<runtime>
-		<assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-			<dependentAssembly>
-				<assemblyIdentity name="Newtonsoft.Json" publicKeyToken="30AD4FE6B2A6AEED" culture="neutral"/>
-				<bindingRedirect oldVersion="0.0.0.0-6.0.0.0" newVersion="6.0.0.0"/>
-			</dependentAssembly>
-		</assemblyBinding>
-	</runtime>
-</configuration>
