--- DocumentDbRepo46/App.config
@@ -1,16 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6" />
-    </startup>
-  <appSettings configSource="private.config">
-  </appSettings>
-  <!--from portal
-  <add key="DocDb:endpoint" value=""/>
-  <add key="DocDb:authkey" value=""/>
-
-  default
-  <add key="DocDb:databaseName" value=""/>
-  <add key="DocDb:collectionName" value=""/>-->
-
-</configuration>
