--- DocumentDbRepo46/Program.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DocumentDbRepo46
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-        }
-    }
-}
