--- DocumentDbRepo46/ScampUser.cs
@@ -1,16 +1,0 @@
-﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DocumentDbRepositories
-{
-    public class ScampUser
-    {
-        [JsonProperty(PropertyName = "id")]
-        public string Id { get; set; }
-        public string Name { get; set; }
-    }
-}
