--- DocumentDbRepo46/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Azure.Documents.Client" version="0.9.2-preview" targetFramework="net46" userInstalled="true" />
-  <package id="Newtonsoft.Json" version="4.5.11" targetFramework="net46" userInstalled="true" />
-</packages>
