+++ ProvisioningJobConsole/project.json
@@ -1,0 +1,16 @@
+{
+    "version": "1.0.0-*",
+    "dependencies": {
+        "Microsoft.Azure.WebJobs": "1.0.1",
+        "Microsoft.Framework.ConfigurationModel.Json": "1.0.0-beta3",
+        "WindowsAzure.Storage": "4.3.0",
+        "Microsoft.Azure.WebJobs.Core": "1.0.1",
+        "ProvisioningLibrary": "1.0.0-*"
+    },
+    "commands": {
+        "ProvisioningJobConsole": "run"
+    },
+    "frameworks": {
+        "aspnet50": { }
+    }
+}
