+++ ProvisioningLibrary/Configuration.cs
@@ -1,0 +1,20 @@
+﻿using System;
+
+namespace ProvisioningLibrary
+{
+    public static class ProvisioningLibraryConfiguration
+    {
+
+
+        public static string GetStorageSubscriptionCertificate()
+        {
+            //This is temporary unitll we get from DB
+            return "MIIKRAIBAzCCCgQGCSqGSIb3DQEHAaCCCfUEggnxMIIJ7TCCBe4GCSqGSIb3DQEHAaCCBd8EggXbMIIF1zCCBdMGCyqGSIb3DQEMCgECoIIE7jCCBOowHAYKKoZIhvcNAQwBAzAOBAjpInwc/qOdzwICB9AEggTIe1mFHCLtWCEXsGFRdeQyoPPZMI20bTA39xg/zp5S2p1Fwc77yu7VLdyq8E3yP3s9R+vQ+n5TyP8LWbQR5lgLQuBnewuphiKl5kTgZYY+ADDcKUJK6lp2mtt0d1SIaRexu+jJHLUM69wOB33J2SWQVR4rWzRKyNuwI2dKDEZZmFrTnmlq7fL5+COwyDbiFHJdHRyJ1a0WmhkJ2nnx68NFEuh3OqHZHHx7m9RqiIIkaKzSUZzbFgm3IcR4F9yUF/4xaddBEPThvmfRb8Hm1XcwAHGxarV447e/T1m75NzmTyi4ZAbzEwy6e2/iGIwnvjov4iPI3aw8g8z5T5aquvARlbZl5kEpW4R3yrSkEkDLt3ldi/oXLHlhaJah36aZTNz/fl00MTDWl/XEI1bbfVmAmT9MofKiJt61CQrYzzLp8NlK0PqOkHJNkkHD6i1bsCNfHwvXrHySo15AV1ivaxEraiOtDq6OulBZj7rvPOYWK3BvdAgJKMHb5dGVi9fN5o4LHXHpwEpnwJG3YcjWfxgmOK0pWIJoVQmG/uAzvGKaxc7aYaNfHeoDZ8us0vBLIo4H6R7xEkmJGnkj3+H3WaMEJ0udl08cch9glTIrTUzcDUR4NsN6uNNu/8mVekTPxN19MZMHrBFQ2UwGJo1MoAwc/St4sYmjY23AF7N/iKdYPaajZc8SwQ19/SAtqdBVE+qMoa+C1Vd1nRDl8WExf4VMq8d3PAbv4wpWxtXTAt5iQy5p6LIxXEOk8KwKyKM77g9XMgdb/4pcPlQZZoUBTirHTg10iDEay75GehEzw7O1In/fD7xDgcbliynvybSOO8KfUajjdX4O4JNVpY9odTpVIV0JFpez3YHInBKvsTf6XXPBMOa67RoSMRbuxthnuXuxpz7TRpeH4j2w4BBrOzLbKJ4XLNTK6YaZ5CXlYYOU2zdhFOCC59sCZewc6eEh0F784/GhIPmHXQidV/fYy68tUxzxIug1DdT/X0vrou2igb4w1J6/jKrneirhvZOVuyEog3dt/R9DIA+uvm/F09GL9SiZYjfM25AzCJw/87+RrlVCLJg2bm0zHetev2MhXoJPihKuntrN2jn0ns8cr9VWYp/fal/7ExR01M+T+CTbPacpjs+hvoYuAft/vlBxAYfpCnH4ilowv5Kkg3+WZ95+1aoCX0pI650I2v+a/gvRUaH7YSV4S4cK9ahvTi/ZMalVIS38RpoysGFbpPvALyaIBHxExkJO33CIpGg7XsGIKokrH94aLXgpiAAWgaL1sQD2uM0kMNCgdiXA7LW/RhbSIlzpi622mN4uqKzIl5FjvhSeRjv+aLNeVzFP2gKFDkskgEqZvKZ+qNawJprn2+pRtTpCGuvOHUo+K45Z0TSSq+n/vsE+bKJkMsX7YAITTMXD9uFSWo1B3e/lZlQpr5ukViBLtevkq5eINvcn2FVCFvOSEQu5zIBmZRoZV/KmEviRJQcra8kPA9VefZxPago8RYAGBgLTKUfkwOW3W+o6GkTLn5rpygFKn85soqV4XN+Ew1D31sijFJFzr3HjKJPy8U4sanLL+J0+7mDFreNn1apAbxm/8nRh+jBPtO0LLZYNku+TuCSz3c+GVE5M2NV6Rif87QAZpWPyMYHRMBMGCSqGSIb3DQEJFTEGBAQBAAAAMFsGCSqGSIb3DQEJFDFOHkwAewBDADcAMAA5ADgAOABBADMALQAyADMANAA5AC0ANAA3AEUAMgAtAEEANABFADgALQBBAEQAMABCAEEAQgA4AEQAMQA2ADEAOAB9MF0GCSsGAQQBgjcRATFQHk4ATQBpAGMAcgBvAHMAbwBmAHQAIABTAG8AZgB0AHcAYQByAGUAIABLAGUAeQAgAFMAdABvAHIAYQBnAGUAIABQAHIAbwB2AGkAZABlAHIwggP3BgkqhkiG9w0BBwagggPoMIID5AIBADCCA90GCSqGSIb3DQEHATAcBgoqhkiG9w0BDAEGMA4ECF20SJBFIktjAgIH0ICCA7D9zMGgVeofW9h5+ivd0eiMTiAjoCle48e/pKubN2OuaIzfsJn61fjIsy+RybZcxsHAP/pfrOGl7+ADweXawLtiYzpq7pCn04Zfkv4Lu9b3h4+6s12+H4EPfG060q7yNRWXo/9TsPtM8HTQDN/rsDPCsEfFzlZUureDqbfIXXg1JWl3uS75VP+PGm/fES6cinnCtdWciyt9JUzaOj3BR+pgvtOypTUOEYPE/PPqsAs6wxWIfaJ8iTBmY6dFVMPEE89//1V80XDkMy7RlXs1zzlkc8kBw8+BUbQjTC0f2EGHkNejh+GrfvgCy6BUkC2Vcw/BuzpCq/4tMppmcAiEx+keIKitptNN7z/mGvLTTpILf1EnycSTrTflEDmSCQmpprt81lx/vKhnJf19Xf6wICfirR4sQrbnaSjO3J7Md7xdDMclyy9TXSnXgTsqg5LVdw1FtudJnKPdDRdIhKC3/KSvw8LxlXEGfBoT+dz+QyG9yO5Bjeba9cwbllakuSCkcQ6JXu9C+nnZ/pM+cLu0RpBeXJ0yV9UhP5kzNR/fsHn1tHAZdLVDIVrz42c5TpdnnGSfA7baCzQsq2tjbKYH+4xzlWGfLulEO3Be/HHL9YnZzRv8bceKJF0jD2uAmgIZIWkef4uvA5VgLcpLPO12nJ81zUJ0eJQke1a1ig7/INeYkla2fYGK0XXSvnDf/h1whv7ySjECl3I7btGm1U+pgr4blOPC4NgeuVh4FTnGSY5BY1+ZlmcZMtehxayMwfs8yjqOV14kMaMCyAQ4peA2Ib+gE/lfgbOw371s0D+gAKvQfrG0/sGV73htxbsCt3V7eDeqWs/8ZS9EAFImj+C0t/N+xc7/q38HxBG+g17Wvj/xZp0D1O/UCscNoYxPe21yu/iRIsERasRdWxc5MIJlbENYiEnee8V1oB4o7hV2G9lI3K9sQl5mk70rEeTcYX0dyDmIumWlXu/odziX616JXenobNqN8Oc8cZJNAZuzaBuI8C8U1B8KlLAKQDFpRtFZ/KkZAarKFJz8B938CXwS3E7HMLTDaE+WpBo9raE+ImoUsgx2CbrataC+v8OSdtfjEZgSLW0Tmyyclpt21si/tWCM1pFk1EMLAlj7ERxeLMLloEB0K91AT4Bxh9totcX4mAnseri/uXnRL6eSBrGnfYcjIh0YAfJq6oh/CLd6bH/zoVEJf63/1D134Ie95SfyDPDD77xDOaDOMTKu3H8f/Li9nUpiVjUFMHTB8Cq9VoFE4TA3MB8wBwYFKw4DAhoEFLNg9EUT/EUmPzl4cMxHHUWOi0U2BBTQYhCoFnxMxn/vitaSPc4/3YOO5Q==";
+        }
+        public static string GetStorageSubscriptionId()
+        {
+            //This is temporary unitll we get from DB
+            return "6210e8d3-3181-4858-89fd-05370bb582d6";
+        }
+    }
+}
