+++ ProvisioningLibrary/QueueMessage.cs
@@ -1,10 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using ProvisioningLibrary;
 
-namespace AzureProvisioningLibrary
+namespace ProvisioningLibrary
 {
     public class QueueMessage
     {
