+++ ProvisioningLibrary/VirtualMachineEnums.cs
@@ -4,9 +4,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AzureProvisioningLibrary
+namespace ProvisioningLibrary
 {
-    
     public enum VirtualMachineAction {Stop=0, Start = 1};
     public enum ResourceAction {Undefined=-1, Stop = 0, Start = 1 , Create =2};
 }
