+++ ProvisioningLibrary/project.json
@@ -1,0 +1,16 @@
+{
+    "version": "1.0.0-*",
+    "dependencies": {
+        "Microsoft.WindowsAzure.Management.Libraries": "2.0.0",
+        "Microsoft.IdentityModel.Clients.ActiveDirectory": "2.14.201151115",
+        "WindowsAzure.Storage": "4.3.0",
+        "DocumentDbRepositories": "1.0.0-*"
+    },
+
+    "frameworks": {
+        "aspnet50": {
+            "dependencies": {
+            }
+        }
+    }
+} }
