+++ ScampApi.sln
@@ -9,16 +9,16 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DataTier", "DataTier", "{FD
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "DocumentDbRepositories", "DocumentDbRepositories\DocumentDbRepositories.kproj", "{0C8E6A9D-D14F-4A5B-9228-09F826D74604}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AzureProvisioningLibrary", "AzureProvisioningLibrary\AzureProvisioningLibrary.csproj", "{0B50A9D4-FEF4-4514-A722-EBF272CCECEB}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{63EC0B60-3D54-4B7D-8BEE-FD226D655CD0}"
 	ProjectSection(SolutionItems) = preProject
 		.nuget\packages.config = .nuget\packages.config
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AzureProvisioningJob", "AzureProvisioningJob\AzureProvisioningJob.csproj", "{1074CC3F-F227-4B44-B95C-AE79D0F8445A}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Provisioning", "Provisioning", "{80B65567-31FB-417F-BF5A-E4D24CA13903}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DocumentDbRepo46", "DocumentDbRepo46\DocumentDbRepo46.csproj", "{ADB3860C-4104-4AB5-BFE7-6C32BBF4657F}"
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "ProvisioningLibrary", "ProvisioningLibrary\ProvisioningLibrary.kproj", "{2C51CD0C-EB46-40FF-BDD4-1F5E8A55B182}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "ProvisioningJobConsole", "ProvisioningJobConsole\ProvisioningJobConsole.kproj", "{BEC8AAD5-37A6-47BE-861A-972FD86E7D18}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -34,24 +34,21 @@ Global
 		{0C8E6A9D-D14F-4A5B-9228-09F826D74604}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0C8E6A9D-D14F-4A5B-9228-09F826D74604}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0C8E6A9D-D14F-4A5B-9228-09F826D74604}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0B50A9D4-FEF4-4514-A722-EBF272CCECEB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0B50A9D4-FEF4-4514-A722-EBF272CCECEB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0B50A9D4-FEF4-4514-A722-EBF272CCECEB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0B50A9D4-FEF4-4514-A722-EBF272CCECEB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1074CC3F-F227-4B44-B95C-AE79D0F8445A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1074CC3F-F227-4B44-B95C-AE79D0F8445A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1074CC3F-F227-4B44-B95C-AE79D0F8445A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1074CC3F-F227-4B44-B95C-AE79D0F8445A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{ADB3860C-4104-4AB5-BFE7-6C32BBF4657F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ADB3860C-4104-4AB5-BFE7-6C32BBF4657F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ADB3860C-4104-4AB5-BFE7-6C32BBF4657F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ADB3860C-4104-4AB5-BFE7-6C32BBF4657F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2C51CD0C-EB46-40FF-BDD4-1F5E8A55B182}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2C51CD0C-EB46-40FF-BDD4-1F5E8A55B182}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2C51CD0C-EB46-40FF-BDD4-1F5E8A55B182}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2C51CD0C-EB46-40FF-BDD4-1F5E8A55B182}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BEC8AAD5-37A6-47BE-861A-972FD86E7D18}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BEC8AAD5-37A6-47BE-861A-972FD86E7D18}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BEC8AAD5-37A6-47BE-861A-972FD86E7D18}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BEC8AAD5-37A6-47BE-861A-972FD86E7D18}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{0C8E6A9D-D14F-4A5B-9228-09F826D74604} = {FD068F2A-2241-4948-A1F7-C756BC85DFAF}
-		{ADB3860C-4104-4AB5-BFE7-6C32BBF4657F} = {FD068F2A-2241-4948-A1F7-C756BC85DFAF}
+		{2C51CD0C-EB46-40FF-BDD4-1F5E8A55B182} = {80B65567-31FB-417F-BF5A-E4D24CA13903}
+		{BEC8AAD5-37A6-47BE-861A-972FD86E7D18} = {80B65567-31FB-417F-BF5A-E4D24CA13903}
 	EndGlobalSection
 EndGlobal
